Remove private header dependancy from public header
